class-33/demos/snacks-from-scratch/pyproject.toml
[tool.poetry]
name = "snacks-from-scratch"
version = "0.1.0"
description = ""
authors = ["JB-Tellez <jb.tellez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework-simplejwt = "^4.6.0"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
Django = "^3.1.4"
djangorestframework = "^3.12.2"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
